% --- Source file: plotData.Rd ---
\name{plotData}
\alias{plotData}
\alias{plotData}

\title{
    Example data for plotting
}
\description{
    A data frame with columns Rank, Type, and MSErr
}
%\details{}
\examples{

    data(plotData, package="SUITOR")

    plotData

}
\keyword{data}
\seealso{ \code{\link{suitor}} }
