% --- Source file: results.Rd ---
\name{results}
\alias{results}
\alias{results}

\title{
    suitor return object
}
\description{
    An object returned from the suitor function for examples
}
%\details{ }
\examples{

    data(results, package="SUITOR")

    results

}
\keyword{data}
\seealso{ \code{\link{suitor}} }
