% --- Source file: data.Rd ---
\name{SimData}
\alias{SimData}
\alias{SimData}

\title{
  Data for examples
}
\description{
 Example input data and results
}
\details{ 
Contains an example input data object of size 96 by 300. 
It is generated by \code{rpois} with mean WH where W (96 by 8) is profile of 8 signatures 
(SBS 4, 6, 7a, 9, 17b, 22, 26, 39) obtained from 
https://cancer.sanger.ac.uk/cosmic/signatures/SBS and H (8 by 300) is rounded integer 
generated from a uniform distribution between 0 and 100 with some randomly 
selected cells being set to zero.
}
\examples{

 data(SimData, package="SUITOR")

 # Display a subset of data objects
 SimData[1:5, 1:5]

}
\keyword{data}
\seealso{ \code{\link{suitor}} }



